+++ src/main/java/com/guitarshack/ReorderThreshold.java
@@ -1,5 +1,5 @@
 package com.guitarshack;
 
 public interface ReorderThreshold {
-    int calculate(Product product);
+    double calculate(Product product);
 }
